Updated service interfaces and implementations

+++ src/main/java/com/repositories/ICustomerRepository.java
@@ -2,9 +2,8 @@ package com.repositories;
 
 import com.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ICustomerRepository extends JpaRepository<Customer, Integer>{
+public interface ICustomerRepository extends JpaRepository<Customer, Long> {
 }
